refactor: move language-ja specific test in language-modules subproject
