+++ language-modules/build.gradle
@@ -1,4 +1,4 @@
-def plugins = [
+def moduleList = [
         [ code: 'ar', lang: 'Arabic'],
         [ code: 'ast', lang: 'Asturian'],
         [ code: 'be', lang: 'Belarusian'],
@@ -31,7 +31,7 @@ def plugins = [
         [ code: 'zh', lang: 'Chinese']
 ]
 
-plugins.forEach { args ->
+moduleList.forEach { args ->
     def name = args.get('code')
     def lang = args.get('lang')
     def pluginClass = "org.omegat.languages.${name}.${lang}Plugin"
@@ -129,4 +129,23 @@ plugins.forEach { args ->
         group = 'other'
     }
     tasks.getByName("checkstyleMain").dependsOn tasks.named("checkstyle${capitalName}")
+}
+
+dependencies {
+    testImplementation(libs.junit4)
+    testRuntimeOnly(libs.slf4j.jdk14)
+    testImplementation(project.rootProject)
+
+    testCompileOnly(dependencies.variantOf(libs.lucene.gosen) { classifier("ipadic") })
+    testImplementation(libs.languagetool.core) {
+        exclude module: 'language-all'
+    }
+    testRuntimeOnly(libs.languagetool.ja) {
+        // Temporary exclusion; see https://sourceforge.net/p/omegat/bugs/814/
+        exclude module: 'lucene-gosen'
+        exclude module: 'languagetool-core'
+        exclude module: 'icu4j'
+    }
+    testRuntimeOnly(dependencies.variantOf(libs.lucene.gosen) { classifier("ipadic") })
+    testRuntimeOnly(libs.icj4j)
 }
