+++ language-modules/src/test/java/org/omegat/languages/test/LuceneGosenCompatibilityTest.java
@@ -23,7 +23,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.omegat.languagetools;
+package org.omegat.languages.test;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -34,15 +34,27 @@ import java.util.Objects;
 
 import net.java.sen.SenFactory;
 import net.java.sen.StringTagger;
-
+import org.junit.BeforeClass;
 import org.junit.Test;
-
 import org.languagetool.JLanguageTool;
-import org.languagetool.Languages;
+import org.languagetool.Language;
 import org.languagetool.rules.RuleMatch;
 import org.languagetool.rules.patterns.PatternRule;
 
+import org.omegat.languagetools.LanguageDataBroker;
+import org.omegat.languagetools.LanguageManager;
+
 public class LuceneGosenCompatibilityTest {
+
+    private static final String JAPANESE = "org.languagetool.language.Japanese";
+
+    @BeforeClass
+    public static void setUpClass() {
+        // We don't use plugin loader in a test context
+        // JLanguageTool.setClassBrokerBroker(new LanguageClassBroker());
+        JLanguageTool.setDataBroker(new LanguageDataBroker());
+        LanguageManager.registerLTLanguage("ja-JP", JAPANESE);
+    }
 
     /**
      * Regression test for bugs#1204.
@@ -57,7 +69,8 @@ public class LuceneGosenCompatibilityTest {
 
     @Test
     public void testJapanese() throws Exception {
-        JLanguageTool lt = new JLanguageTool(Objects.requireNonNull(Languages.getLanguageForName("Japanese")));
+        Language lang = LanguageManager.getLTLanguage(new org.omegat.util.Language("ja-JP"));
+        JLanguageTool lt = new JLanguageTool(Objects.requireNonNull(lang));
         List<RuleMatch> matches = lt.check("そんじゃそこらのやつらとは違う");
         assertEquals(1, matches.size());
         assertTrue(matches.get(0).getRule() instanceof PatternRule);
